Separate empty container class.
Contributes to CURA-5628.
